+++ src/singlerender.cpp
@@ -1,9 +1,9 @@
 #include "renderers.h"
 #include <QPainter>
 
-void drawSingle(const QImage &img, int panX, int panY, QPainter &painter, float zoom){
-    if(zoom <= 0.0f){
-        zoom = qMin(float(painter.window().width()) / float(img.width()), float(painter.window().height()) / float(img.height()));
+void drawSingle(const QImage &img, int panX, int panY, QPainter &painter, qreal zoom){
+    if(zoom <= 0.0){
+        zoom = qMin(qreal(painter.window().width()) / qreal(img.width()), qreal(painter.window().height()) / qreal(img.height()));
     }
     painter.translate(painter.window().width() / 2,
                       painter.window().height() / 2);
